refactor(utils): move process_worker + card_images_workers into image_service
Both modules are imported only from services/image_service/. Co-locate them
as internal infrastructure of the service; rename card_images_workers.py to
workers.py since it's now scoped inside image_service.

Refs #441
