+++ services/image_service/process_worker.py
