+++ services/image_service/workers.py
